fix(morphology): geodetic length/thickness from real perimeter (was size_t-truncated)
geo_len_thickness.cpp assigned PERIMETER (a double) into a size_t and then used integer
division, truncating the perimeter to its floor and rounding every intermediate term. On
the 8x8 shape2d fixture this made GEODETIC_LENGTH 10.0 and THICKNESS 3.0 -- ~11% off the
rectangle-model formula (geodetic ~11.13, thickness ~2.34). Keep perimeter and area as
double and use floating-point division, so both features equal the documented side lengths
of the rectangle with the same area and perimeter (the roots of x^2 - (P/2)x + A = 0).

Vet both by documented-formula conformance (like CIRCULARITY / DIAMETER_EQUAL_PERIMETER):
GEODETIC_LENGTH == P/4 + sqrt(P^2/16 - A), THICKNESS == P/2 - GEODETIC_LENGTH, recomputed
at 1e-9 from the vetted AREA + PERIMETER in
test_shape2d_documented_formula_conformance_no_external_oracle. GEODETIC_LENGTH
regression -> vetted (analytic); THICKNESS imea -> analytic (its prior imea row was
tracker-only -- test_morphology_imea.h never existed -- and the loose 1000x snapshot had
pinned the truncated 3.0). Goldens updated; the redundant loose geodetic gtest removed.
morphology 80 -> 81.

gtest 716/716; pytest 72 passed (7 arrow fails are the tiff-only local build; no Python
test covers these features). Coverage 607 -> 608 / 758.

+++ src/nyx/features/geo_len_thickness.cpp
@@ -16,20 +16,23 @@ GeodeticLengthThicknessFeature::GeodeticLengthThicknessFeature() : FeatureMethod
 
 void GeodeticLengthThicknessFeature::calculate (LR& r, const Fsettings& s)
 {
-	size_t roiArea = r.aux_area,
+	// Keep perimeter as a real number: assigning it into size_t truncated the fractional part and
+	// forced integer division below, so geodetic_length/thickness deviated from the rectangle-model
+	// formula (e.g. ~11% on the 8x8 shape2d fixture: geodetic 10.0 instead of ~11.13).
+	double roiArea = r.aux_area,
 		roiPerimeter = r.fvals[(int)Feature2D::PERIMETER][0];
 
-	double SqRootTmp = roiPerimeter * roiPerimeter / 16 - (double)roiArea;
+	double SqRootTmp = roiPerimeter * roiPerimeter / 16.0 - roiArea;
 
 	// Make sure value under SqRootTmp is always positive
-	if (SqRootTmp < 0) 
+	if (SqRootTmp < 0)
 		SqRootTmp = 0;
 
 	// Calculate geodetic_length with pq-formula (see above):
-	geodetic_length = roiPerimeter / 4 + sqrt(SqRootTmp);
+	geodetic_length = roiPerimeter / 4.0 + sqrt(SqRootTmp);
 
 	// Calculate thickness by rewriting Equation (2):
-	thickness = roiPerimeter / 2 - geodetic_length;
+	thickness = roiPerimeter / 2.0 - geodetic_length;
 }
 
 void GeodeticLengthThicknessFeature::osized_calculate (LR& r, const Fsettings& s, ImageLoader&)
